カスタムフックに書き換え
